refactor(flame): extract BubbleManager from TechWorld god object
Extract ~730 lines of bubble lifecycle management into a dedicated
BubbleManager class, reducing TechWorld from 2081 to 1392 lines (-33%).

BubbleManager is a plain Dart class (not a Flame Component) that owns:
- Proximity detection and bubble creation/removal
- Physics repulsion between overlapping bubbles
- Metaball field and merged video rendering
- Audio enable/disable based on Chebyshev distance
- Shader loading (video bubble, metaball, merged video)
- Dreamfinder avatar bridge lifecycle

It receives an addComponent callback and shared map references to
remotePlayers/bots, keeping component tree ownership in TechWorld.

Adds 21 unit tests for chebyshevDistance, update lifecycle, audio
threshold, clear/dispose, and all public API edge cases.
